10k Processing updates
