--- data/costco_financial_data.csv
@@ -1,26 +1,0 @@
-item,fiscal_year,value,unit
-Earnings Per Share Diluted,2024,16.56,per share
-Earnings Per Share Diluted,2023,14.16,per share
-Earnings Per Share Diluted,549,249.0,per share
-Gross Margin,2024,249625.0,millions
-Gross Margin,549,806.0,per share
-Membership Fee Revenue,2024,249625.0,millions
-Membership Fee Revenue,2021,689.0,per share
-Membership Fee Revenue,549,806.0,per share
-Merchandise Costs,2024,222358.0,millions
-Merchandise Costs,2023,212586.0,millions
-Merchandise Costs,2022,199382.0,millions
-Net Income,2024,7367.0,millions
-Net Income,2023,6292.0,millions
-Net Income,549,249625.0,millions
-Net Sales,2024,60.0,per share
-Net Sales,2023,237710.0,millions
-Net Sales,2022,222730.0,millions
-Net Sales,549,249625.0,millions
-Operating Income,2024,7367.0,millions
-Operating Income,2023,165294.0,millions
-Operating Income,549,806.0,per share
-Total Revenue,2024,184143.0,millions
-Total Revenue,2023,176630.0,millions
-Total Revenue,2022,165294.0,millions
-Total Revenue,549,7367.0,millions
