--- data/extract_financial_tables.py
@@ -1,323 +1,0 @@
-#!/usr/bin/env python3
-"""
-Extract financial tables from Costco 10-K narrative text.
-Creates a structured database and lookup tool for financial data.
-"""
-
-import re
-import sqlite3
-import pandas as pd
-from pathlib import Path
-from typing import List, Dict, Optional, Tuple
-import json
-
-class FinancialTableExtractor:
-    def __init__(self):
-        self.data_dir = Path(__file__).parent
-        self.narrative_file = self.data_dir / "costco_narrative.txt"
-        self.db_file = self.data_dir / "costco_financial_data.db"
-        self.csv_file = self.data_dir / "costco_financial_data.csv"
-        
-    def extract_tables(self):
-        """Extract financial tables from narrative text."""
-        with open(self.narrative_file, 'r', encoding='utf-8') as f:
-            content = f.read()
-            lines = content.splitlines()
-        
-        # Find and extract financial data
-        financial_data = []
-        
-        # Extract specific financial items using patterns
-        patterns = [
-            # Net sales pattern: "Net sales increased X% to $XXX,XXX"
-            (r'Net sales.*?\$\s*([\d,]+)', 'Net Sales', 2024),
-            # Look for tabular data
-            (r'(\d{4})\s*\n.*?Net [Ss]ales\s*\n\$\s*([\d,]+)', 'Net Sales', None),
-            (r'(\d{4})\s*\n.*?Membership fee.*?\n\$\s*([\d,]+)', 'Membership Fee Revenue', None),
-            (r'(\d{4})\s*\n.*?Total revenue.*?\n\$\s*([\d,]+)', 'Total Revenue', None),
-            (r'(\d{4})\s*\n.*?Cost of goods sold.*?\n\$\s*([\d,]+)', 'Cost of Goods Sold', None),
-            (r'(\d{4})\s*\n.*?Gross margin.*?\n\$\s*([\d,]+)', 'Gross Margin', None),
-            (r'(\d{4})\s*\n.*?Operating income.*?\n\$\s*([\d,]+)', 'Operating Income', None),
-            (r'(\d{4})\s*\n.*?Net income.*?\n\$\s*([\d,]+)', 'Net Income', None),
-            # Per share data
-            (r'(\d{4})\s*\n.*?diluted.*?\n\$\s*([\d.]+)', 'Earnings Per Share Diluted', None),
-        ]
-        
-        # Search for patterns
-        for pattern, item_name, default_year in patterns:
-            matches = re.finditer(pattern, content, re.IGNORECASE | re.MULTILINE | re.DOTALL)
-            for match in matches:
-                if default_year:
-                    year = default_year
-                    value_str = match.group(1)
-                else:
-                    year = int(match.group(1))
-                    value_str = match.group(2)
-                
-                # Clean and convert value
-                value_str = value_str.replace(',', '')
-                try:
-                    value = float(value_str)
-                    financial_data.append({
-                        'item': item_name,
-                        'fiscal_year': year,
-                        'value': value,
-                        'unit': 'millions' if value > 1000 else 'per share'
-                    })
-                except ValueError:
-                    pass
-        
-        # Manual extraction of key data from the narrative
-        # Based on the grep results, we know these values:
-        manual_data = [
-            {'item': 'Net Sales', 'fiscal_year': 2024, 'value': 249625, 'unit': 'millions'},
-            {'item': 'Net Sales', 'fiscal_year': 2023, 'value': 237710, 'unit': 'millions'},
-            {'item': 'Net Sales', 'fiscal_year': 2022, 'value': 222730, 'unit': 'millions'},
-            {'item': 'Membership Fee Revenue', 'fiscal_year': 2024, 'value': 4828, 'unit': 'millions'},
-            {'item': 'Total Revenue', 'fiscal_year': 2024, 'value': 254453, 'unit': 'millions'},
-            {'item': 'Total Revenue', 'fiscal_year': 2023, 'value': 242290, 'unit': 'millions'},
-            {'item': 'Total Revenue', 'fiscal_year': 2022, 'value': 226954, 'unit': 'millions'},
-            {'item': 'Merchandise Costs', 'fiscal_year': 2024, 'value': 222358, 'unit': 'millions'},
-            {'item': 'Merchandise Costs', 'fiscal_year': 2023, 'value': 212586, 'unit': 'millions'},
-            {'item': 'Merchandise Costs', 'fiscal_year': 2022, 'value': 199382, 'unit': 'millions'},
-            {'item': 'Gross Margin', 'fiscal_year': 2024, 'value': 27267, 'unit': 'millions'},
-            {'item': 'Net Income', 'fiscal_year': 2024, 'value': 7367, 'unit': 'millions'},
-            {'item': 'Net Income', 'fiscal_year': 2023, 'value': 6292, 'unit': 'millions'},
-            {'item': 'Earnings Per Share Diluted', 'fiscal_year': 2024, 'value': 16.56, 'unit': 'per share'},
-            {'item': 'Earnings Per Share Diluted', 'fiscal_year': 2023, 'value': 14.16, 'unit': 'per share'},
-        ]
-        
-        # Add manual data
-        financial_data.extend(manual_data)
-        
-        # Remove duplicates
-        seen = set()
-        unique_data = []
-        for item in financial_data:
-            key = (item['item'], item['fiscal_year'])
-            if key not in seen:
-                seen.add(key)
-                unique_data.append(item)
-        
-        # Create DataFrame
-        df = pd.DataFrame(unique_data)
-        df = df.sort_values(['item', 'fiscal_year'], ascending=[True, False])
-        
-        # Save to CSV
-        df.to_csv(self.csv_file, index=False)
-        print(f"Financial data saved to CSV: {self.csv_file}")
-        
-        # Save to SQLite
-        self._create_database(df)
-        
-        return df
-    
-    def _create_database(self, df: pd.DataFrame):
-        """Create SQLite database with financial data."""
-        conn = sqlite3.connect(self.db_file)
-        
-        # Create main table
-        df.to_sql('financial_data', conn, if_exists='replace', index=False)
-        
-        # Create indexes
-        conn.execute('CREATE INDEX IF NOT EXISTS idx_item ON financial_data(item)')
-        conn.execute('CREATE INDEX IF NOT EXISTS idx_year ON financial_data(fiscal_year)')
-        
-        # Create view for easy querying
-        conn.execute('''
-            CREATE VIEW IF NOT EXISTS financial_summary AS
-            SELECT 
-                item,
-                fiscal_year,
-                value,
-                unit,
-                CASE 
-                    WHEN unit = 'millions' THEN value * 1000000
-                    ELSE value
-                END as actual_value
-            FROM financial_data
-            ORDER BY item, fiscal_year DESC
-        ''')
-        
-        conn.commit()
-        conn.close()
-        
-        print(f"Financial data saved to SQLite: {self.db_file}")
-        print(f"Total records: {len(df)}")
-        
-        # Display sample
-        print("\nSample financial data:")
-        print(df.head(10))
-
-
-class StructuredDataLookup:
-    """Tool for querying structured financial data without using LLM."""
-    
-    def __init__(self, db_path: str = None):
-        if db_path is None:
-            db_path = Path(__file__).parent / "costco_financial_data.db"
-        self.db_path = db_path
-        
-        # Query patterns
-        self.patterns = {
-            'net sales': 'Net Sales',
-            'revenue': 'Total Revenue',
-            'sales': 'Net Sales',
-            'income': 'Net Income',
-            'earnings': 'Net Income',
-            'eps': 'Earnings Per Share Diluted',
-            'earnings per share': 'Earnings Per Share Diluted',
-            'cost': 'Merchandise Costs',
-            'cogs': 'Merchandise Costs',
-            'cost of goods': 'Merchandise Costs',
-            'gross margin': 'Gross Margin',
-            'gross profit': 'Gross Margin',
-            'membership': 'Membership Fee Revenue',
-            'membership fee': 'Membership Fee Revenue',
-        }
-    
-    def structured_data_lookup(self, query: str) -> Dict:
-        """
-        Query financial data using natural language.
-        
-        Args:
-            query: Natural language query like "Net sales 2024"
-            
-        Returns:
-            Dictionary with query results
-        """
-        query_lower = query.lower()
-        
-        # Extract year
-        year_match = re.search(r'20\d{2}', query)
-        year = int(year_match.group()) if year_match else None
-        
-        # Find item to search
-        item_to_search = None
-        for pattern, item_name in self.patterns.items():
-            if pattern in query_lower:
-                item_to_search = item_name
-                break
-        
-        # If no pattern matched, try to find closest match
-        if not item_to_search:
-            # Look for capitalized words
-            words = [w for w in query.split() if w[0].isupper() and not w.isdigit()]
-            if words:
-                potential_item = ' '.join(words)
-                # Check if it's close to any known items
-                for pattern, item_name in self.patterns.items():
-                    if pattern in potential_item.lower():
-                        item_to_search = item_name
-                        break
-        
-        if not item_to_search:
-            return {
-                'query': query,
-                'error': 'Could not identify financial item in query',
-                'available_items': list(set(self.patterns.values()))
-            }
-        
-        # Query database
-        results = self._query_db(item_to_search, year)
-        
-        return {
-            'query': query,
-            'item_searched': item_to_search,
-            'year_filter': year,
-            'results': results
-        }
-    
-    def _query_db(self, item: str, year: Optional[int]) -> List[Dict]:
-        """Query the database."""
-        if not Path(self.db_path).exists():
-            return [{'error': f'Database not found at {self.db_path}'}]
-        
-        conn = sqlite3.connect(self.db_path)
-        conn.row_factory = sqlite3.Row
-        
-        try:
-            if year:
-                query = '''
-                    SELECT item, fiscal_year, value, unit, actual_value
-                    FROM financial_summary
-                    WHERE item = ? AND fiscal_year = ?
-                '''
-                cursor = conn.execute(query, (item, year))
-            else:
-                query = '''
-                    SELECT item, fiscal_year, value, unit, actual_value
-                    FROM financial_summary
-                    WHERE item = ?
-                    ORDER BY fiscal_year DESC
-                    LIMIT 3
-                '''
-                cursor = conn.execute(query, (item,))
-            
-            results = []
-            for row in cursor:
-                results.append({
-                    'item': row['item'],
-                    'fiscal_year': row['fiscal_year'],
-                    'value': row['value'],
-                    'unit': row['unit'],
-                    'actual_value': row['actual_value'],
-                    'formatted_value': f"${row['value']:,.0f} {row['unit']}" if row['unit'] == 'millions' else f"${row['value']:.2f}"
-                })
-            
-            return results
-            
-        except Exception as e:
-            return [{'error': str(e)}]
-        finally:
-            conn.close()
-    
-    def get_available_items(self) -> List[str]:
-        """Get list of available financial items."""
-        if not Path(self.db_path).exists():
-            return []
-        
-        conn = sqlite3.connect(self.db_path)
-        cursor = conn.execute('SELECT DISTINCT item FROM financial_data ORDER BY item')
-        items = [row[0] for row in cursor]
-        conn.close()
-        return items
-
-
-def main():
-    """Extract financial data and test lookup."""
-    # Extract financial tables
-    extractor = FinancialTableExtractor()
-    df = extractor.extract_tables()
-    
-    print("\n" + "="*50)
-    print("Testing Structured Data Lookup")
-    print("="*50)
-    
-    # Test lookup
-    lookup = StructuredDataLookup()
-    
-    # Show available items
-    print("\nAvailable financial items:")
-    for item in lookup.get_available_items():
-        print(f"  - {item}")
-    
-    # Test queries
-    test_queries = [
-        "Net sales 2024",
-        "Total revenue 2023",
-        "Net income 2024",
-        "EPS 2024",
-        "Gross margin",
-        "Membership fee revenue 2024"
-    ]
-    
-    print("\nTest queries:")
-    for query in test_queries:
-        print(f"\nQuery: '{query}'")
-        result = lookup.structured_data_lookup(query)
-        print(f"Result: {json.dumps(result, indent=2)}")
-
-
-if __name__ == "__main__":
-    main()
